fix: reset event selection upon starting to create a new event
